Add mirage deployment pipeline, add mirage bootstrap options
