+++ .gitmodules
@@ -1,7 +1,7 @@
 [submodule "helper-scripts"]
 	path = helper-scripts
 	url = https://github.com/skalenetwork/helper-scripts.git
-	branch = develop
+	branch = add-mirage-container
 [submodule "hardhat-node"]
 	path = hardhat-node
 	url = git@github.com:skalenetwork/hardhat-node.git
